[PR 10921] [peripherals] Fix 100% cpu usage after a while
When float is large enough, small increments no longer increase it.

+++ xbmc/peripherals/Peripherals.h
@@ -214,7 +214,7 @@ namespace PERIPHERALS
      * @brief rateHz The rate in Hz
      * @return A handle that unsets its rate when expired
      */
-    EventRateHandle SetEventScanRate(float rateHz) { return m_eventScanner.SetRate(rateHz); }
+    EventRateHandle SetEventScanRate(double rateHz) { return m_eventScanner.SetRate(rateHz); }
 
     /*!
      *
